--- lib/cln_plugin.dart
@@ -1,6 +1,0 @@
-/// This is a template for plugins written in
-/// the dart programming environment.
-
-library cln_plugin;
-
-export 'src/cln_plugin.dart';
