+++ lib/kraken.dart
@@ -1,0 +1,6 @@
+/// This is a template for plugins written in
+/// the dart programming environment.
+
+library kraken;
+
+export 'src/kraken.dart';
